Website Frontend
JS + CSS files have now been updated in folder, as well as the main home page html. This includes the fade in up and the navigation bar at the top. Still need to implement the bottom sections of the page and mess with the colors, but front end looks nice.